add user movie service impl

+++ src/main/kotlin/ru/melowetty/filmswishlistservice/controller/request/MovieUserRateRequest.kt
@@ -6,5 +6,5 @@ import jakarta.validation.constraints.Min
 data class MovieUserRateRequest(
     @Min(1, message = "{movie.rate.request.min-rate}")
     @Max(10, message = "{movie.rate.request.max-rate}")
-    val rate: Float
+    val rate: Int
 )
